Added support for dry-run

+++ devopstemplate/template.py
@@ -15,24 +15,24 @@ import devopstemplate.pkg as pkg
 class DevOpsTemplate():
 
     def __init__(self, projectdirectory='.',
-                 overwrite_exists=False, skip_exists=False):
-        logger = logging.getLogger('DevOpsTemplate.__init__')
-        if not os.path.exists(projectdirectory):
-            logger.info('creating project directory: %s', projectdirectory)
-            os.makedirs(projectdirectory)
+                 overwrite_exists=False, skip_exists=False, dry_run=False):
+        # logger = logging.getLogger('DevOpsTemplate.__init__')
         self.__projectdir = projectdirectory
         self.__overwrite = overwrite_exists
         self.__skip = skip_exists
-        # template_index_list = pkg.string_list('template.index')
-        # logger.debug('template index:\n%s', '\n'.join(template_index_list))
+        self.__dry_run = dry_run
+        with pkg.stream('template.json') as fh:
+            self.__template_dict = json.load(fh)
 
         self.__env = Environment(loader=PackageLoader(__name__, 'template'),
                                  autoescape=select_autoescape(default=True))
+        # Create project base directory if not present
+        self.__mkdir(projectdirectory)
 
-    def create(self, projectname, projectconfig):
+    def create(self, projectconfig):
         logger = logging.getLogger('DevOpsTemplate.create')
         logger.info('Create project from template')
-        logger.info('Project name: %s', projectname)
+        logger.info('Project name: %s', projectconfig['project_name'])
         logger.debug('  scripts directory: %s',
                      projectconfig['add_scripts_dir'])
         logger.debug('  docs directory:    %s',
@@ -45,6 +45,12 @@ class DevOpsTemplate():
                      not projectconfig['no_sonar'])
         logger.debug('  Docker support:    %s',
                      not projectconfig['no_docker'])
+
+        # Install 'common' files
+        common_file_list = self.__template_dict['common']
+        for template_fpath in common_file_list:
+            project_fpath = Template(template_fpath).render(**projectconfig)
+            self.__copy(template_fpath, project_fpath)
 
     def manage(self, add_gitignore, add_readme, add_scripts,
                add_docs, add_sonar, add_docker):
@@ -60,7 +66,8 @@ class DevOpsTemplate():
         project_dpath = os.path.join(self.__projectdir, project_dname)
         if not os.path.exists(project_dpath):
             logger.info('creating directory: %s', project_dpath)
-            os.makedirs(project_dpath)
+            if not self.__dry_run:
+                os.makedirs(project_dpath)
         else:
             logger.debug('directory %s exists', project_dpath)
 
@@ -87,9 +94,15 @@ class DevOpsTemplate():
             raise FileExistsError(f'File {project_fpath} already exists, exit.'
                                   ' (use --skip-exists or --overwrite-exists'
                                   ' to control behavior)')
-        with pkg.stream(pkg_fname) as pkg_fh:
-            with open(project_fpath, 'wb') as project_fh:
-                shutil.copyfileobj(pkg_fh, project_fh)
+        if not self.__dry_run:
+            # Create parent directories if not present
+            parent_dname = os.path.dirname(project_fpath)
+            if not os.path.exists(parent_dname):
+                os.makedirs(parent_dname)
+            # Copy file in binary mode
+            with pkg.stream(pkg_fname) as pkg_fh:
+                with open(project_fpath, 'wb') as project_fh:
+                    shutil.copyfileobj(pkg_fh, project_fh)
         logger.debug('template:%s  ->  project:%s', pkg_fname, project_fpath)
 
     def __render_file(self):
